REFACTOR :: 폴더 구조, 파일 명, 코드 컨벤션
