+++ src/app/layout.tsx
@@ -6,11 +6,11 @@ export const metadata = {
   description: "부산소프트웨어마이스터고등학교 입학전형 시스템 마루입니다",
 };
 
-interface PropsInterface {
+interface PropsType {
   children: ReactNode;
 }
 
-const RootLayout = ({ children }: PropsInterface) => {
+const RootLayout = ({ children }: PropsType) => {
   return (
     <html lang="en">
       <body>{children}</body>
