+++ src/components/Common/Input/type.ts
@@ -1,0 +1,12 @@
+import { InputHTMLAttributes } from "react";
+
+export interface InputPropsType extends InputHTMLAttributes<HTMLInputElement> {
+  width?: string;
+  desc?: string;
+  msg?: string;
+}
+
+export interface ButtonInputPropsType extends InputPropsType {
+  buttonText: string;
+  buttonClick: () => void;
+}
